Install all components
Before this commit, inside a `describe` speclj would only install the
component returned by each inner expression. This moves installation to
the place where components are created, so that test code like this
works as expected and runs 2 tests instead of just 1:

(let [x 1]
  (it "works 1" (= x 1))
  (it "works 2" (= x 1)))

This has always been a bit of a stumbling block, and while we'd normally
use an `around` or similar to contextualize code, there's no reason I
can see that this shouldn't just work.

refs #163