Manually reverted 42dbd8db (PR #164).
Unfortunately, this commit hampered running specs in ClojureScript
(`lein cljs`). The high-level descriptions were being found (and
reported) but none of their contained characteristics or nested contexts
were executed.

As a result of this revert, the new 'focus'-related features are
somewhat hampered. Specifically, a context that is nested within a
focused context is not being executed as expected. I hope to resolve
this in a subsequent commit.